+++ Source/BetaHubBugReporter/Private/BH_PluginSettings.cpp
@@ -13,6 +13,8 @@ UBH_PluginSettings::UBH_PluginSettings()
     ShortcutKey = EKeys::F12;
     MaxRecordedFrames = 30;
     MaxRecordingDuration = 60;
+    MaxVideoWidth = 1920;
+    MaxVideoHeight = 1080;
 
     static ConstructorHelpers::FClassFinder<UBH_ReportFormWidget> WidgetClassFinder1(TEXT("/BetaHubBugReporter/BugReportForm"));
     static ConstructorHelpers::FClassFinder<UBH_PopupWidget> WidgetClassFinder2(TEXT("/BetaHubBugReporter/BugReportFormPopup"));